Update 3.2.3.3.1 Accounting Requirements
We do not need currently an accounting office because we do not participate as an economic entity and as such the law in force does not apply to our course program. If in the future a new law enters into force, we will change it in accordance with the law.