Update 3.2.3.3.2 Security Requirements
The personal data of all users of this system will be safeguarded in accordance with the existing regulations and legislative standards in Albania. The Commission for Personal Data Protection, located in Albania, grants licenses to institutions, organizations, and corporations that enable them to handle personal and sensitive data pertaining to users of information systems. 

The data protection laws in Albania are primarily governed by the Law No. 9887 "On Personal Data Protection" (known as the "Albanian Data Protection Law"). This law was adopted in 2008 and provides regulations for the collection, processing, storage, and transfer of personal data in Albania.  law grants individuals certain rights, including the right to be informed, the right to access their personal data, the right to rectify or erase their personal data, and the right to object to the processing of their personal data. Any use of personal data for the purpose of damaging the image or sharing personal information is punishable according to the penal code of the Republic of Albania